feat: phase 2 stabilization milestone - visual backends and devguide finalization
- Stabilized molsysviewer.MolSysView and nglview.NGLWidget forms.
- Hardened H5MSM and PDB file handlers with polymorphic instantiation.
- Integrated viewers as active molecular systems in is_a_molecular_system.
- Unified all form internal delegates with @arg_digest for strict compliance.
- Finalized Phase 2 documentation in devguide/ (Maturity Audit, Support Tiers).
- Fixed residual bugs in PDB text conversion and Structure setters.

+++ molsysmt/pbc/get_lengths_and_angles_from_box.py
@@ -27,7 +27,8 @@ def get_lengths_and_angles_from_box(box, skip_digestion=False):
     """
 
     box_value, box_unit  = puw.get_value_and_unit(box)
-    lengths_value, angles_value = msmlib.pbc.get_lengths_and_angles_from_box(box_value)
+    import numpy as np
+    lengths_value, angles_value = msmlib.pbc.get_lengths_and_angles_from_box(box_value.astype(np.float64))
     lengths = puw.quantity(lengths_value.round(6), box_unit)
     lengths = puw.standardize(lengths)
     angles = puw.quantity(angles_value.round(6), 'radians')
